+++ database/migrations/2024_01_01_000018_create_purchases_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->foreignId('inventory_id')->nullable()->constrained()->nullOnDelete();
             $table->string('reference');
             $table->enum('status', PurchaseStatus::keys())->default(PurchaseStatus::default()->value);
-            $table->unsignedInteger('total_amount');
-            $table->unsignedInteger('net_amount')->default(0);
-            $table->unsignedInteger('paid_amount')->default(0);
+            $table->decimal('total_amount', 15, 2);
+            $table->decimal('net_amount', 15, 2)->default(0);
+            $table->decimal('paid_amount', 15, 2)->default(0);
             $table->string('note')->nullable();
             $table->foreignId('created_by')->nullable()->constrained('users')->nullOnDelete();
             $table->foreignId('updated_by')->nullable()->constrained('users')->nullOnDelete();
