+++ database/migrations/2026_08_15_000001_create_purchase_refunds_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('purchase_id')->constrained();
             $table->foreignId('purchase_return_id')->constrained();
-            $table->unsignedInteger('amount');
+            $table->decimal('amount', 15, 2);
             $table->string('status')->default(PurchaseRefundStatus::PAID->value);
             $table->foreignId('created_by')->nullable()->constrained('users')->nullOnDelete();
             $table->foreignId('updated_by')->nullable()->constrained('users')->nullOnDelete();
